apps/mcp-server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "MCP Server for UIT AI Assistant - Document retrieval and DAA scraping tools"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=0.1.0",
    "llama-index-core>=0.11.0",
    "llama-index-vector-stores-chroma>=0.1.0",
    "llama-index-embeddings-openai>=0.1.0",
    "llama-index-llms-openai>=0.1.0",
    "llama-index-llms-gemini>=0.1.0",
    "llama-index-llms-ollama>=0.1.0",
    "chromadb>=0.5.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "playwright>=1.40.0",
    "crawl4ai>=0.1.0",
    "FlagEmbedding>=1.2.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "modal>=0.60.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
